refactor and move to api dir
